backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sentence-transformers==3.0.0
faiss-cpu==1.8.0
google-generativeai==0.7.2
pdfplumber==0.11.0
python-docx==1.1.0
supabase==2.5.0
python-multipart
python-jose[cryptography]
python-dotenv